Have Swift methods call NetHack's C panic function, instead of making our own. Other minor code improvements.

+++ win/nh3d/winnh3d.m
@@ -1393,7 +1393,7 @@ You("スコアの載らない発見モードで起動した．");
 
 - (void)didPresentError:(NSError *)error
 {
-	int result;
+	NSInteger result;
 	NSAlert *alert = [ NSAlert alertWithError:error ];
 	result = [ alert runModal ];
 }
